+++ packages/visual-stack/src/components/Drawer/index.js
@@ -1,0 +1,15 @@
+import './styles.css';
+
+import classNames from 'classnames';
+import React from 'react';
+
+export const Drawer = ({ className, children, isOpen, ...restProps }) => (
+  <div
+    {...restProps}
+    className={classNames(className, 'vs-drawer', {
+      'vs-drawer-open': isOpen,
+    })}
+  >
+    {children}
+  </div>
+);